Load sandboxes in the sidenav
