+++ src/app/api/api.module.ts
@@ -8,6 +8,7 @@ import { ServicesResolver } from './resolvers/services-resolver.service';
 import { PreferencesStorage } from './services/preferences-storage.service';
 import { DebugStatusApi } from './services/debug-status-api.service';
 import { AuthService } from './services/auth.service';
+import { SandboxeService } from './services/sandboxe.service';
 
 @NgModule({
   imports: [CommonModule],
@@ -19,6 +20,7 @@ import { AuthService } from './services/auth.service';
     PreferencesStorage,
     DebugStatusApi,
     AuthService,
+    SandboxeService,
   ],
 })
 export class ApiModule {}
