+++ src/app/api/interfaces/sandboxe.interface.ts
@@ -1,0 +1,4 @@
+export interface Sandbox {
+  businessId: string;
+  name?: string;
+}
