+++ src/app/app.component.ts
@@ -1,8 +1,10 @@
-import { Component } from '@angular/core';
+import { Component, OnInit } from '@angular/core';
 
 import { Angulartics2GoogleTagManager } from 'angulartics2/gtm';
 import { AuthService } from './api/services/auth.service';
 import { Observable } from 'rxjs/Observable';
+import { SandboxeService } from './api/services/sandboxe.service';
+import { Sandbox } from './api/interfaces/sandboxe.interface';
 
 @Component({
   selector: 'zp-root',
@@ -13,22 +15,31 @@ import { Observable } from 'rxjs/Observable';
       </zp-header>
 
       <zp-sidenav
-        [toggle]="toggledSidenav">
+        [toggle]="toggledSidenav"
+        [sandboxes]="this.sandboxes">
       </zp-sidenav>
     </div>
 
     <router-outlet name="login"></router-outlet>
   `,
 })
-export class AppComponent {
+export class AppComponent implements OnInit {
   isLoggedIn$: Observable<boolean>;
   toggledSidenav = true;
+  sandboxes: Array<Sandbox> = [];
 
   constructor(
     authService: AuthService,
     analytics: Angulartics2GoogleTagManager,
+    private sandboxeService: SandboxeService,
   ) {
     this.isLoggedIn$ = authService.isLoggedIn();
+  }
+
+  ngOnInit() {
+    this.sandboxeService.getSandboxes().then(sandboxes => {
+      this.sandboxes = sandboxes;
+    });
   }
 
   onToggledSidenav(event: boolean) {
