+++ src/app/shared/sidenav/sidenav.component.scss
@@ -23,8 +23,7 @@
       }
     }
 
-    .room-list,
-    .user-list {
+    .sandboxe-list {
       margin-bottom: 15px;
 
       mat-list-item {
