+++ src/app/terminal/terminal-view/terminal-view.component.html
@@ -1,6 +1,6 @@
 <div class="terminal-container">
   <h1>
-    <mat-icon class="back-link" mat-list-icon [routerLink]="['/sandboxes']">arrow_back</mat-icon>
+    <mat-icon class="back-link" mat-list-icon [routerLink]="['/dashboard']">arrow_back</mat-icon>
     <span>Terminal for {{ sandboxId }}</span>
     <a mat-raised-button [routerLink]="['/traces', sandboxId]" color="primary"><mat-icon>folder</mat-icon> Table View</a>
   </h1>
