+++ src/app/traces/traces-view/traces-view.component.html
@@ -1,6 +1,6 @@
 <mat-sidenav-container class="Container--Sidenav">
   <h1>
-    <mat-icon mat-list-icon [routerLink]="['/sandboxes']">arrow_back</mat-icon>
+    <mat-icon mat-list-icon [routerLink]="['/dashboard']">arrow_back</mat-icon>
     <span>Trace for {{sandboxId}}</span>
     <a mat-raised-button [routerLink]="['/terminal', sandboxId]" color="primary"><mat-icon>subtitles</mat-icon> Terminal View</a>
   </h1>
